Move Dynamo back to core. Performance goes all weird
