+++ Source/Core/CMakeLists.txt
@@ -3,6 +3,7 @@ add_library(Core OBJECT
 
                         CPU.cpp 
                         DMA.cpp 
+                        Dynamo.cpp
                         FlashMem.cpp  
                         Interpret.cpp 
                         Interrupt.cpp
