+++ Source/Core/Dynamo.cpp
@@ -21,7 +21,7 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 #include <fstream>
 
 #include "Base/Types.h"
-#include "DynaRec/Dynamo.h"
+#include "Core/Dynamo.h"
