+++ Source/Core/Dynamo.h
